Spliting the bookmarks_menu_and_toolbar.py file in smaller tests
